+++ packages/waallet/background/provider.ts
@@ -1,6 +1,6 @@
 import type { AccountManager } from "~packages/account/manager"
-import { BundlerRpcMethod } from "~packages/bundler/rpc"
-import { GasPriceEstimator } from "~packages/gas/price/estimator"
+import { GasPriceEstimator } from "~packages/eip/4337/bundler/estimator"
+import { BundlerRpcMethod } from "~packages/eip/4337/bundler/rpc"
 import type { NetworkManager } from "~packages/network/manager"
 import { JsonRpcProvider } from "~packages/rpc/json/provider"
 import address from "~packages/util/address"
